+++ packages/itwin/measure-tools/src/api/GraphicStyle.ts
@@ -3,9 +3,10 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 
-import { Arc3d, LineSegment3d, Path, Point3d, XYProps } from "@itwin/core-geometry";
-import { ColorByName, ColorDef, ColorDefProps, LinePixels } from "@itwin/core-common";
-import { GraphicBuilder } from "@itwin/core-frontend";
+import type { Arc3d, LineSegment3d, Path, Point3d, XYProps } from "@itwin/core-geometry";
+import type { ColorDefProps, LinePixels } from "@itwin/core-common";
+import { ColorByName, ColorDef } from "@itwin/core-common";
+import type { GraphicBuilder } from "@itwin/core-frontend";
 
 /** List of well-known graphic styles. */
 export enum WellKnownGraphicStyleType {