+++ packages/itwin/measure-tools/src/tools/MeasurePerpendicularTool.ts
@@ -3,15 +3,20 @@
 * See LICENSE.md in the project root for license terms and full copyright notice.
 *--------------------------------------------------------------------------------------------*/
 
-import { Plane3dByOriginAndUnitNormal, Point3d, Ray3d, Vector3d } from "@itwin/core-geometry";
-import {
-  BeButtonEvent, EventHandled, IModelApp, OutputMessagePriority, TentativeOrAccuSnap, ToolAssistance, ToolAssistanceImage, ToolAssistanceInputMethod,
+import type { Vector3d } from "@itwin/core-geometry";
+import { Plane3dByOriginAndUnitNormal, Point3d, Ray3d } from "@itwin/core-geometry";
+import type {
+  BeButtonEvent,
   ToolAssistanceInstruction, ToolAssistanceSection,
 } from "@itwin/core-frontend";
-import { Feature, MeasureToolsFeatures } from "../api/FeatureTracking";
+import {
+  EventHandled, IModelApp, OutputMessagePriority, TentativeOrAccuSnap, ToolAssistance, ToolAssistanceImage, ToolAssistanceInputMethod,
+} from "@itwin/core-frontend";
+import type { Feature } from "../api/FeatureTracking";
+import { MeasureToolsFeatures } from "../api/FeatureTracking";
 import { MeasurementToolBase } from "../api/MeasurementTool";
 import { MeasurementViewTarget } from "../api/MeasurementViewTarget";
-import { DistanceMeasurement } from "../measurements/DistanceMeasurement";
+import type { DistanceMeasurement } from "../measurements/DistanceMeasurement";
 import { MeasureDistanceToolModel } from "../toolmodels/MeasureDistanceToolModel";
 
 export class MeasurePerpendicularTool extends MeasurementToolBase<DistanceMeasurement, MeasureDistanceToolModel> {